terrorctf/website/requirements.txt
Flask==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
